Refactor modelos de Expense e Income para usar int en lugar de BigDecimal para el monto; Y pruebas unitarias para los controladores.

+++ src/main/java/com/jaime/finance_springboot_app/models/Expense.java
@@ -1,6 +1,5 @@
 package com.jaime.finance_springboot_app.models;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import jakarta.persistence.Entity;
@@ -26,7 +25,7 @@ public class Expense {
     private Long id;
 
     private String description;
-    private BigDecimal amount;
+    private int amount;
     private LocalDate date;
