+++ src/main/java/com/jaime/finance_springboot_app/models/Income.java
@@ -1,6 +1,5 @@
 package com.jaime.finance_springboot_app.models;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import jakarta.persistence.Entity;
@@ -25,7 +24,7 @@ public class Income {
     private Long id;
 
     private String description;
-    private BigDecimal amount;
+    private int amount;
     private LocalDate date;
 
     @ManyToOne
